Add smoothing algorithm for color histogram
In order to pick the 'best color range', we'll need to smooth the colors
produced by the histogram function. This is a combination of numpy's convolve()
and a smoothing function, like hamming or bartlett.

This includes a demonstration that plots the various algorithms so we can talk
about them during our color analysis section.