Share idea for generating random test sequence
I’ve created a function that generates a random number sequence that shaped
_somewhat_ like a bell curve. Let’s put something like this in the tests
directory where we can import it whenever we want to play around with some of
the standard deviations and other functions.

Like all the things I am sharing, it is just to iterate specific helper
functions we've talked about.