Abstraction for NetworkTables. (#14)
* Add a robot reader and writer

* Got nt woking with epoch and 'persistent = True'

* add pyfrc to pipfiles

* Add instructions for installing network tables instructions to README

* Including network tables dependency.

* Add networktables code from last year

The 'lib/tables.py' is a simple interface to properly connecting to a
NetworkTables server. We should probably add a good default for the
server, e.g. 10.27.33.2 or 127.0.0.1

* network table abstraction

* Add abstraction to the network tables interface.

README has instructions for downloading and running a network table server.
has_network_tables.py checks that the network tables server works.
tables.py contains an abstraction that sends values to the network tables.
It is not final.

* Delete code involving epochs