Raspberry Pi Interface and Documentation
All files concerning the raspberry pi have been moved to a folder. The README has been extensively
documented with where to move these files and other commands to set up the raspberry pi.
Includes a program for the pi to connect to wifi, and when intially powered on downloads all neccessary
libraries and runs robot_vision immediately.