Addressing Delay (#27)
* Addressing Delay

Removed many keys that were being sent to network tables. Added global variable of frame to calculate the input for the motors
(a range from 1 to -1) rather than having that processing done on the roborio.

* Addressing Delay

Removed many keys that were being sent to network tables. Added global variable of frame to calculate the input for the motors
(a range from 1 to -1) rather than having that processing done on the roborio.